requirements.txt
python-dotenv
textblob
requests
